move run-semgrep script into composite action to facilitate calling workflow cross-repo
