+++ .github/workflows/run_semgrep_scan.yml
@@ -99,34 +99,9 @@ jobs:
           # Full history only when diff/baseline is requested
           fetch-depth: ${{ inputs.semgrep_scan_mode == 'full' && '1' || '0' }}
 
-      - name: Extract workflow definition repo and ref
-        id: extract_workflow_details
-        run: |
-          WF_REF="${{ github.workflow_ref }}"
-          echo "github.workflow_ref: $WF_REF"
-          REPO="${WF_REF%%/.github/workflows/*}"
-          REF="${WF_REF##*@}"
-          echo "Extracted repo: $REPO"
-          echo "Extracted ref: $REF"
-          echo "repo=$REPO" >> $GITHUB_OUTPUT
-          echo "ref=$REF" >> $GITHUB_OUTPUT
-
-      - name: Checkout workflow repo for access to internal scripts/actions
-        uses: actions/checkout@v4
-        with:
-          repository: ${{ steps.extract_workflow_details.outputs.repo }}
-          ref: ${{ steps.extract_workflow_details.outputs.ref }}
-          path: workflow-repo
-
-      - run: |
-          ls
-          echo "-----"
-          cd workflow-repo
-          ls
-
       - name: Check for open PR (by commit)
         id: pr_check
-        uses: ./workflow-repo/.github/actions/pr-open-check
+        uses: OpenSesame/core-github-actions/.github/actions/pr-open-check@actions/pr-open-check/2.0.0
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           commit-identifier: ${{ inputs.commit_identifier }}
@@ -151,6 +126,7 @@ jobs:
 
       - name: Run Semgrep
         id: semgrep
+        uses: OpenSesame/core-github-actions/.github/actions/run-semgrep@actions/run-semgrep/1.0.0
         env:
           INPUT_BASELINE: ${{ inputs.baseline_ref }}
           HAS_PR: ${{ steps.pr_check.outputs.pr_exists }}
@@ -160,7 +136,6 @@ jobs:
           SEMGREP_TARGETS: ${{ inputs.semgrep_targets }}
           FAIL_LEVEL: ${{ inputs.fail_severity }}
           EXTRA_ARGS: ${{ inputs.extra_args }}
-        run: node ./workflow-repo/scripts/gha-lib/run-semgrep.js
 
       - name: Upload Artifact
         if: ${{ steps.semgrep.outputs.totalFindings > 0 }}
@@ -290,7 +265,7 @@ jobs:
 
       - name: Upsert PR comment
         if: ${{ github.event_name == 'pull_request' || steps.pr_check.outputs.pr_exists == 'true' }}
-        uses: ./workflow-repo/.github/actions/upsert-pr-comment
+        uses: OpenSesame/core-github-actions/.github/actions/upsert-pr-comment@actions/upsert-pr-comment/1.0.0
         with:
           pr-number: ${{ steps.pr_check.outputs.pr_number }}
           github-token: ${{ secrets.GITHUB_TOKEN }}
